+++ src/net.h
@@ -278,6 +278,9 @@ public:
     bool DisconnectNode(const std::string& node);
     bool DisconnectNode(NodeId id);
 
+    bool SetWhitelisted(NodeId id, bool fWhitelisted);
+    bool SetManualConnection(NodeId id, bool manual);
+
     ServiceFlags GetLocalServices() const;
 
     //!set the max outbound target in bytes