+++ src/rpc/net.cpp
@@ -135,8 +135,8 @@ static UniValue getpeerinfo(const JSONRPCRequest& request)
 {
     if (request.fHelp || request.params.size() > 1)
         throw std::runtime_error(
-            "getpeerinfo\n"
-            "\nReturns data about connected network nodes as a json array of objects.\n"
+            "getpeerinfo \"id\"\n"
+            "\nReturns data about connected network nodes as a json array of objects\n"
             "\nArguments:\n"
             "1. \"id\"         (numeric, optional) Only return information about the peer with the specified id.\n"
             "\nResult:\n"
@@ -214,6 +214,55 @@ static UniValue getpeerinfo(const JSONRPCRequest& request)
     }
 
     return ret;
+}
+
+static UniValue updatepeer(const JSONRPCRequest& request)
+{
+    if (request.fHelp || request.params.size() < 1 || request.params.size() > 3)
+        throw std::runtime_error(
+            "updatepeer\n"
+            "\nUpdate settings for a network peer (debug method).\n"
+            "\nReturns updated information about the peer. See the help text for getpeerinfo for the format of the returned object."
+            "\nArguments:\n"
+            "1. \"id\"                (string, required) The peer id (see getpeerinfo for a list of peers with their ids)\n"
+            "2. \"whitelisted\"       (bool, optional) whether the peer is whitelisted\n"
+            "3. \"manual_connection\" (bool, optional) whether the peer was manually added\n"
+            "\nExamples:\n"
+            + HelpExampleCli("updatepeer", "0 true")
+            + HelpExampleRpc("updatepeer", "0, true")
+        );
+
+    if (!g_connman) {
+        throw JSONRPCError(RPC_CLIENT_P2P_DISABLED, "Error: Peer-to-peer functionality missing or disabled");
+    }
+
+    RPCTypeCheckArgument(request.params[0], UniValue::VNUM);
+    int node_id = request.params[0].get_int();
+
+    CNodeStats stats;
+    if (!g_connman->GetNodeStats(node_id, stats)) {
+        throw JSONRPCError(RPC_INVALID_ADDRESS_OR_KEY, strprintf("Node id %d not found", node_id));
+    }
+
+    if (!request.params[1].isNull()) {
+        RPCTypeCheckArgument(request.params[1], UniValue::VBOOL);
+        if (!g_connman->SetWhitelisted(node_id, request.params[1].isTrue())) {
+            throw JSONRPCError(RPC_MISC_ERROR, "Failed to update peer whitelisting");
+        }
+    }
+
+    if (!request.params[2].isNull()) {
+        RPCTypeCheckArgument(request.params[2], UniValue::VBOOL);
+        if (!g_connman->SetManualConnection(node_id, request.params[2].isTrue())) {
+            throw JSONRPCError(RPC_MISC_ERROR, "Failed to update peer manual_connection");
+        }
+    }
+
+    g_connman->GetNodeStats(node_id, stats);
+
+    UniValue entry(UniValue::VOBJ);
+    NodeStatsToJSON(stats, entry);
+    return entry;
 }
 
 static UniValue addnode(const JSONRPCRequest& request)
@@ -665,6 +714,7 @@ static const CRPCCommand commands[] =
     { "network",            "listbanned",             &listbanned,             {} },
     { "network",            "clearbanned",            &clearbanned,            {} },
     { "network",            "setnetworkactive",       &setnetworkactive,       {"state"} },
+    { "hidden",             "updatepeer",             &updatepeer,             {"id", "whitelisted", "manual_connection"} },
 };
 
 void RegisterNetRPCCommands(CRPCTable &t)